+++ doc/CHANGES.rst
@@ -27,6 +27,8 @@ CHANGELOG
 * Fixed #227 str and repr of the query call str query.sql_statement
 * Refactored main methods to get the description of the models and fields
   without used the existing table
+* Added new column type **ModelSelection** this role is to replace a column
+  with a foreigh key to the deprecated model **Model.System.Model**
 
 2.0.0 (2023-03-11)
 ------------------